refactor, data exploration, papertype class, biorxiv_all file cannot be added, add notes
