+++ .gitignore
@@ -8,4 +8,4 @@ keywords/tests/__pycache__
 paperscraper/.ipynb_checkpoints/
 paperscraper/check_coverage.ipynb
 paperscraper/testfile.ipynb
-
+data/biorxiv_all*.jsonl
